+++ 2dgs/FileManagement/BodyData.cs
@@ -7,6 +7,6 @@ public class BodyData
     public string Name;
     public Vector2 Position { get; set; }
     public Vector2 Velocity { get; set; }
-    public int Mass { get; set; }
+    public float Mass { get; set; }
     public float DisplayRadius { get; set; }
 }
